requirements.txt
discord.py>=1.3
python-dotenv==0.13.0